refactor: documentStorageSchema into internal
